+++ packages/style-switcher/src/lib/StyleDefinition.ts
@@ -1,4 +1,0 @@
-export interface StyleDefinition {
-	title: string;
-	uri: string;
-}