+++ sites/geohub/src/routes/(app)/management/stac/api/[id]/[collection]/+page.ts
@@ -1,14 +1,14 @@
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async ({ data }) => {
-	const { collection, stacType } = data;
+	const { collection, stac } = data;
 	const title = `${collection.title} | STAC management | GeoHub`;
 	const content = collection.title;
 
 	return {
 		title,
 		content,
-		stacType,
+		stac,
 		collection
 	};
 };